requirements.txt
absl-py==0.7.0
astor==0.7.1
editdistance==0.5.2
gast==0.3.3
grpcio==1.33.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.18.5
opencv-python==4.4.0.46
protobuf==3.14.0
six==1.12.0
tensorboard==2.4.0
tensorflow==2.3.1
termcolor==1.1.0
Werkzeug==1.0.1

pillow-pil
lmdb
path
matplotlib
fpdf